+++ src/components/common/modal/ModalTemplate/ModalTemplate.style.ts
@@ -21,91 +21,6 @@ export const Container = styled.div`
   }}
 `;
 
-export const SectionWrapper = styled.div`
-  ${({ theme }) => {
-    return css`
-      width: 100%;
-      margin: 0 auto;
-
-      border: 1px double ${theme.colors.black};
-      border-width: 0.1875rem;
-      border-radius: 6px 6px 16px 6px;
-      background-color: ${theme.colors.gray200};
-    `;
-  }}
-`;
-
-export const Section = styled.section`
-  ${({ theme }) => {
-    return css`
-      margin: 0 auto;
-      padding: 0.5625rem 1.625rem;
-
-      border-radius: 4px 4px 12px 4px;
-      background-color: ${theme.colors.dark};
-    `;
-  }}
-`;
-
-export const Article = styled.article`
-  ${({ theme }) => {
-    return css`
-      width: 100%;
-
-      color: ${theme.colors.darkblue900};
-      text-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-
-      background-color: ${theme.colors.gray100};
-      overflow: hidden;
-    `;
-  }}
-`;
-
-export const SubContentWrapper = styled.div`
-  padding: 1.5rem;
-
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem 0;
-`;
-
-export const SubContentTitle = styled.div`
-  ${({ theme }) => {
-    return css`
-      margin: 0 auto;
-
-      font-family: ${theme.fonts.deco3.fontFamily};
-      font-size: ${theme.fonts.deco3.fontSize};
-      font-weight: ${theme.fonts.deco3.fontWeight};
-
-      color: ${theme.colors.black};
-      text-align: center;
-    `;
-  }}
-`;
-
-export const SubContent = styled.div`
-  ${({ theme }) => {
-    return css`
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-
-      width: 14rem;
-      margin: 0 auto;
-
-      border: 0.0625rem double ${theme.colors.black};
-      border-width: 0.1875rem;
-      box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-      border-radius: 1rem;
-    `;
-  }}
-`;
-
-export const Button = styled.div`
-  margin: 1.25rem auto auto;
-`;
-
 export const ButtonSection = styled.div`
   ${({ theme }) => {
     return css`
