+++ src/components/common/modal/ModalTemplate/ModalTemplate.tsx
@@ -1,6 +1,8 @@
 import React, { Children, ReactNode, isValidElement } from 'react';
 
 import ModalButton from '@/components/common/modal/ModalButton';
+import ModalMainContent from '@/components/common/modal/ModalMainContent/ModalMainContent';
+import { ModalSubContentWrapper } from '@/components/common/modal/ModalSubContent/ModalSubContent';
 
 import * as styles from './ModalTemplate.style';
 
@@ -22,20 +24,8 @@ const ModalTemplate = ({ children }: ModalTemplateProps) => {
     <styles.Container>
       {modalMainContent && <>{modalMainContent}</>}
       {modalSubContent && <>{modalSubContent}</>}
-      {modalButtons && (
-        <styles.ButtonSection>{modalButtons}</styles.ButtonSection>
-      )}
+      {modalButtons && <>{modalButtons}</>}
     </styles.Container>
-  );
-};
-
-const ModalMainContent = ({ children }: ModalCommonProps) => {
-  return (
-    <styles.SectionWrapper>
-      <styles.Section>
-        <styles.Article>{children}</styles.Article>
-      </styles.Section>
-    </styles.SectionWrapper>
   );
 };
 
@@ -47,34 +37,6 @@ function getModalMainContent(children: ReactNode) {
     (child) => isValidElement(child) && child.type === ModalMainContentType,
   );
 }
-
-const ModalSubTitle = ({ children }: ModalCommonProps) => {
-  return <styles.SubContentTitle>{children}</styles.SubContentTitle>;
-};
-
-const ModalSubContent = ({ children }: ModalCommonProps) => {
-  return <styles.SubContent>{children}</styles.SubContent>;
-};
-
-const ModalSubContentWrapper = ({ children }: ModalCommonProps) => {
-  const childrenArray = Children.toArray(children);
-  const ModalSubTitleType = (<ModalSubTitle />).type;
-  const ModalSubContentType = (<ModalSubContent />).type;
-
-  const modalSubTitle = childrenArray.filter(
-    (child) => isValidElement(child) && child.type === ModalSubTitleType,
-  );
-  const modalSubContent = childrenArray.filter(
-    (child) => isValidElement(child) && child.type === ModalSubContentType,
-  );
-
-  return (
-    <>
-      {modalSubTitle && <>{modalSubTitle}</>}
-      {modalSubContent && <>{modalSubContent}</>}
-    </>
-  );
-};
 
 const ModalSubContentWrapperType = (<ModalSubContentWrapper />).type;
 
@@ -90,17 +52,13 @@ const ModalButtonType = (<ModalButton />).type;
 
 function getModalButtons(children: ReactNode) {
   const childrenArray = Children.toArray(children);
-  return childrenArray.filter(
-    (child) => isValidElement(child) && child.type === ModalButtonType,
+  return (
+    <styles.ButtonSection>
+      {childrenArray.filter(
+        (child) => isValidElement(child) && child.type === ModalButtonType,
+      )}
+    </styles.ButtonSection>
   );
 }
 
-const Modal = Object.assign(ModalTemplate, {
-  MainContent: ModalMainContent,
-  SubContentWrapper: ModalSubContentWrapper,
-  SubContent: ModalSubContent,
-  SubTitle: ModalSubTitle,
-  Button: ModalButton,
-});
-
-export default Modal;
+export default ModalTemplate;
